Revert "[DSv4] Refactor compressor & Fix ROCm compatibility (#43710)"
This reverts commit adaa5e455ad89525e8afa69b0633ec9c4b720f7e.
